Fix inconsistent font size on mobile
The dropdowns / radio buttons render with a much smaller font size than the intro text. This is because the page will get automatically downscaled on most mobile devices and the scaling algorithm will try to compensate by adjusting the font sizes. Unfortunately, the dropdowns and radio buttons seem to be omitted from this for some reason. This change disables the text scaling altogether, meaning all the text will be equally small.