Fix inconsistent font size on mobile (#48)
* Fix inconsistent font size on mobile

The dropdowns / radio buttons render with a much smaller font size than the intro text. This is because the page will get automatically downscaled on most mobile devices and the scaling algorithm will try to compensate by adjusting the font sizes. Unfortunately, the dropdowns and radio buttons seem to be omitted from this for some reason. This change disables the text scaling altogether, meaning all the text will be equally small.

* More horizontal graph space

Removing the title of the y-axis and reducing the right margin from 80px to 40px to save some horizontal space.

* Fixed content width on mobile

Force 750px page width on mobile devices. This will cause the text to appear slightly larger but the graph will be slightly narrower.